requirements.txt
azure-functions==1.15.0
requests
pyodbc
pandas
python-dotenv
